Dead Space - Impossible NewGame+ 🚀🧑‍🚀👽
I have finally managed to beat the game on Impossible difficulty!
It is an acheievement I had on my personal to do list since the release of the game in 2008
And since the game does not have cloud saving, this is imperative,
that I save a copy of my save-files, feel free to use them yourself!